requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
python-dotenv==1.0.0
aiofiles==23.2.0
pandas==2.1.4
asyncio-throttle==1.0.2
markdown==3.5.2
pytz==2023.4